+++ messages/ru-RU/app/modules/admin.php
@@ -64,6 +64,10 @@ return [
     '&larr; Prev page' => '&larr; Предыдущая страница',
     'Next page &rarr;' => 'Следующая страница &rarr;',
 
+    'Favourites' => "Избранное",
+    'Favourite' => "В избранное",
+    'Un Favourite' => "Удалить",
+
     'Information' => "Информация",
     'System information' => "Информация о системе",
     'IP, host' => "IP, хост",